ssd1306: fixup driver `bug`
Note, this is a super weird issue (likely isolated to the ebay
displays), where the driver IC seems to be an sh1106 despite being
advertised as an ssd1306.

However, the only works consistently with the additions made in this
commit. Still unsure why? checked the register maps, no such cmds
exist.

But the display wont work exclusively in SSD1306_DEFUALT mode either, so
it works with the SH1106 write function... confusion...

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>